refactor(devices): extract proto parsers into devices_parser module
Move the eight pure proto-to-Device parsers (parse_device, _parse_statuses,
_parse_device_state, _parse_battery, _parse_spread_properties,
_parse_hub_device, _parse_video_edge_channel, _dedupe_video_doorbells) plus
their lookup maps and the wire-shape/redaction helpers out of devices.py into a
new side-effect-free api/devices_parser.py.

DevicesApi keeps thin @staticmethod delegators that forward to the module, so
the historical DevicesApi.parse_device / DevicesApi._parse_* call surface used
by the coordinator and the 76 test references is unchanged — zero test churn.
The _handle_update live path imports the shared maps/helpers from the new
module; the two test-imported helpers are re-exported from devices.py. The
parser module pins its logger name to api.devices so debug output (and the
caplog-based tests) keep their existing namespace.

devices.py drops from 1194 to ~590 lines. Pure refactor — no behaviour change,
all 1396 tests pass unchanged, 88% coverage. Tier-2 #2 from the refactor
backlog.
